requirements.txt
distro==1.7.0
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
tabula-py==2.5.0